// src/share/enum/connection.enum.ts
export enum ConnectionNameEnum {
  LOOKUP_TRAFFIC = 'LOOKUP_TRAFFIC',
}

export enum CustomerEmailTypeEnum {
  USERNAME_PASSWORD = 'mail',
  SOCIAL = 'sso',
  ZALO = 'zalo',
  OTHERS = 'others',
  FACEBOOK = 'facebook',
}

export enum GenderEnum {
  MALE = 'male',
  FEMALE = 'female',
  ORTHERS = 'others',
}

export enum AppEnv {
  DEVELOPMENT = 'development',
  PRODUCTION = 'production',
}

export enum TrackingRegisterLogTypeEnum {
  NEW_MEMBER = 'new_member',
  CDS_MEMBER = 'cds_member',
}

export enum MessageResponse {
  SUCCESS = 'Success',
  FAILED = 'Failed',
}